Update spider test and luogu

+++ spider/luogu/luogu.go
@@ -2,9 +2,9 @@ package luogu
 
 import "XCPCer_board/scraper"
 
-func ScrapeAll(uid string) (map[string]int64, error) {
+func ScrapeAll(uid string) (map[string]int, error) {
 	// 请求所有并合并所有
-	res, err := scraper.MergeAllResults[string, int64](
+	res, err := scraper.MergeAllResults[string, int](
 		GetStrMsg(uid),
 	)
 	if err != nil {
